+++ OInjo_Shop/src/main/java/kr/idu/OInjo_Shop/repository/Item/ItemImgRepository.java
@@ -1,9 +1,12 @@
 package kr.idu.OInjo_Shop.repository.Item;
 
 import kr.idu.OInjo_Shop.entity.Item.ItemImgEntity;
-import kr.idu.OInjo_Shop.entity.Item.Relation.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface ItemImgRepository extends JpaRepository<ItemImgEntity, Long> {
+
+    List<ItemImgEntity> findByItem(Long itemId);
 }
